Refactor display rendering to shared library with u32 timestamps
- Remove font-generator crate (replaced by greeting-renderer with Parley)
- Add shared rendering functions to voyager-display (render_info,
  render_scrolling_bmp, clear_display, advance_scroll)
- Change timestamp handling from u64 to u32 throughout to avoid
  AtomicU64 issues on Cortex-M4 (valid until year 2106)
- Update firmware to use shared voyager-display rendering
- Simplify font-viewer to use voyager-display
- Move greetings.txt to greeting-renderer
- Remove outdated README files

+++ shared-icd/src/lib.rs
@@ -7,12 +7,12 @@
 
 use ergot::{endpoint, topic};
 
-// Time service endpoints
-endpoint!(SetTimeEndpoint, u64, (), "time/set");
-endpoint!(GetTimeEndpoint, (), u64, "time/get");
+// Time service endpoints (u32 timestamp valid until year 2106)
+endpoint!(SetTimeEndpoint, u32, (), "time/set");
+endpoint!(GetTimeEndpoint, (), u32, "time/get");
 
 // DFU endpoint
 endpoint!(RebootToDfuEndpoint, (), (), "dfu/reboot");
 
 // Topics for notifications
-topic!(TimeUpdatedTopic, u64, "time/updated");
+topic!(TimeUpdatedTopic, u32, "time/updated");
